% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Celegans_frontal.R
\docType{data}
\name{Celegans.frontal}
\alias{Celegans.frontal}
\title{Load C. elegans data : neuronal network of frontal neurons}
\format{
a named list containing\describe{
\item{igraph}{an \code{'igraph'} object.}
\item{name}{a length-\eqn{131} vector of node names.}
\item{position}{an \eqn{(131\times 2)} matrix of physical positions.}
}
}
\usage{
data(Celegans.frontal)
}
\description{
It is a directed neuronal network of frontal neurons consisting of 131 nodes 
and 764 binary edges. Additionally, two-dimensional coordinates for all nodes 
are given.
}
\examples{
\donttest{
# load the data
data(Celegans.frontal, package="T4network")
}

}
\references{
Kaiser M and Hilgetag CC (2006). "Nonoptimal component placement, but short processing paths, due to long-distance projections in neural systems." \emph{PLoS Comput. Biol.}, 2.7:e95.
}
\concept{data}
\keyword{datasets}
