% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dolphins.R
\docType{data}
\name{dolphins}
\alias{dolphins}
\title{Load dolphins data}
\format{
a named list containing\describe{
\item{igraph}{an \code{'igraph'} object.}
\item{name}{a length-\eqn{62} vector of node names for dolphins.}
\item{label}{a length-\eqn{62} vector of known community membership.}
}
}
\usage{
data(dolphins)
}
\description{
This data consists of social network of 62 dolphins. This small undirected network 
is known to consist of two communities of sizes 41 and 21 nodes each.
}
\examples{
\donttest{
# load the data
data(dolphins, package="T4network")
}

}
\concept{data}
\keyword{datasets}
