% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effective.R
\name{effective}
\alias{effective}
\title{Compute Effective Resistance}
\usage{
effective(graph)
}
\arguments{
\item{graph}{one of the followings; (1) an \code{igraph} object, (2) a \code{network} object, or (3) an \eqn{(N\times N)} adjacency matrix.}
}
\value{
an \eqn{(N\times N)} matrix of pairwise effective resistance values.
}
\description{
This function computes the effective resistance, a concept of dissimilarity for nodes. 
A notable characteristic is that the square root of effective resistance values 
for a given network induces a metric structure of a graph. For simplicity, this function 
only accepts a network of \eqn{\lbrace 0, 1 \rbrace } binary edges.
}
\examples{
## load the karate club data
data(karate, package="T4network")

## compute effective resistance & extract the label
karate.eff = effective(karate$A)
karate.lab = karate$label

## find the metric structure and embed using MDS
karate.met = stats::as.dist(sqrt(karate.eff))
karate.2d  = stats::cmdscale(karate.met)

## visualize
opar <- par(no.readonly=TRUE)
plot(karate.2d, col=karate.lab, pch=19, xlab="x", ylab="y", 
     main="Embedding with MDS+Effective Resistance")
par(opar)

}
\references{
\insertRef{young_new_2016}{T4network}

\insertRef{young_new_2016-1}{T4network}
}
\concept{compute}
