% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_embed2007CYT.R
\name{embed2007CYT}
\alias{embed2007CYT}
\title{Directed Graph Embedding by Chen, Yang, and Tang (2007)}
\usage{
embed2007CYT(graph, ndim = 2, alpha = 0.01)
}
\arguments{
\item{graph}{one of the followings; (1) an \code{igraph} object, (2) a \code{network} object, or (3) an \eqn{(N\times N)} adjacency matrix.}

\item{ndim}{an embedding dimension for a given graph (default: 2).}

\item{alpha}{perturbation factor \eqn{\in (0,1)} (default: 0.01).}
}
\value{
a named list containing\describe{
\item{embed}{an \eqn{(N\times ndim)} matrix of embedded coordinates.}
}
}
\description{
Find a low-dimensional embedding of the network given an asymmetric/directed graph. 
For simplicity, this function only accepts a network of \eqn{\lbrace 0, 1 \rbrace } binary edges.
}
\examples{
\donttest{
## create a simple directed ring graph
library(igraph)
mygraph = graph.ring(10, directed=TRUE)

## embed in R^2
embed2  = embed2007CYT(mygraph)$embed

## visualize the results
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(mygraph, main="directed ring graph")
plot(embed2, pch=19, main="2-dimensional embedding",
     xlab="dimension 1", ylab="dimension 2")
par(opar)
} 

}
\references{
\insertRef{chen_directed_2007}{T4network}
}
\concept{embedding}
