% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gcpPHATE.R
\name{gcpPHATE}
\alias{gcpPHATE}
\title{Community Detection with PHATE : Partitional Approach}
\usage{
gcpPHATE(
  graph,
  k = 2,
  metrize = c("effective", "shortest"),
  nnbd = 5,
  alpha = 2
)
}
\description{
Community Detection with PHATE : Partitional Approach
}
\examples{
## load the data
data(karate, package="T4network")

## community detection & given label
lab.true = karate$label
lab.eff  = gcpPHATE(karate$A, k=2, metrize="effective")$cluster
lab.spd  = gcpPHATE(karate$A, k=2, metrize="shortest")$cluster

## prepare for plotting with 'igraph' with fixed layout
obj.igraph = igraph::graph_from_adjacency_matrix(karate$A, mode="undirected")
obj.igraph$layout <- igraph::layout_with_kk

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3))
plot(obj.igraph, vertex.label=NA, vertex.color=lab.true, main="given label")
plot(obj.igraph, vertex.label=NA, vertex.color=lab.eff,  main="gcPHATE-effective")
plot(obj.igraph, vertex.label=NA, vertex.color=lab.spd,  main="gcPHATE-shortest")
par(opar)

}
\concept{cluster}
