% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_karate.R
\docType{data}
\name{karate}
\alias{karate}
\title{Load Zachary's Karate club network}
\format{
a data.frame containing\describe{
\item{A}{a \eqn{(34\times 34)} adjacency matrix that is binary and symmetric.}
\item{label}{a length-\eqn{34} vector of class membership.}
}
}
\usage{
data(karate)
}
\description{
This is the famous Karate club network data, which contains the adjacency matrix 
and class label only.
}
\examples{
\donttest{
# load the data
data(karate, package="T4network")
}

}
\references{
Zachary WW (1977). "An Information Flow Model for Conflict and Fission in Small Groups." \emph{Journal of Anthropological Research}, \bold{33}(4):452-473.
}
\concept{data}
\keyword{datasets}
