% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lsm.R
\name{lsm}
\alias{lsm}
\title{Fit a Latent Space Model for a Binary Network}
\usage{
lsm(graph, ndim = 2)
}
\description{
Fit a Latent Space Model for a Binary Network
}
\examples{
## load the karate club data
data(karate, package="T4network")

## extract the label
karate.lab = karate$label

## compute 2-dimensional latent space
karate.lsm = lsm(karate$A, ndim=2)

## visualize
opar <- par(no.readonly=TRUE)
plot(karate.lsm$latent, col=karate.lab, pch=19, xlab="x", ylab="y", 
     main="Latent Space Embedding")
par(opar)

}
\concept{inference}
