% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effective.R
\name{effective}
\alias{effective}
\title{Compute Effective Resistance}
\usage{
effective(graph)
}
\arguments{
\item{graph}{one of the followings; (1) an \code{igraph} object, (2) a \code{network} object, or (3) an \eqn{(N\times N)} adjacency matrix.}
}
\value{
an \eqn{(N\times N)} matrix of pairwise effective resistance values.
}
\description{
This function computes the effective resistance, a concept of dissimilarity for nodes. 
A notable characteristic is that the square root of effective resistance values 
for a given network induces a metric structure of a graph. For simplicity, this function 
only accepts a network of \eqn{\lbrace 0, 1 \rbrace } binary edges.
}
\examples{
## load the karate club data
data(karate, package="T4network")

## compute effective resistance & compute metrics
effres = effective(karate$A)
effmet = sqrt(effres)

## visualize
opar <- par(no.readonly=TRUE)
par(pty="s")
image(effmet[,34:1], axes=FALSE, main="metric matrix from ER")
par(opar)

}
\references{
\insertRef{young_new_2016}{T4network}

\insertRef{young_new_2016-1}{T4network}
}
\concept{compute}
