% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lsm.R
\name{lsm}
\alias{lsm}
\title{Fit a Latent Space Model for a Binary Network}
\usage{
lsm(graph, ndim = 2, nstart = 5)
}
\description{
Fit a Latent Space Model for a Binary Network
}
\examples{
\donttest{
## load the karate club data
data(karate, package="T4network")

## compute 2-dimensional latent space
lsm.embed = lsm(karate$A, ndim=2)

## prepare for plotting with 'igraph' with fixed layout
obj.igraph = igraph::graph_from_adjacency_matrix(karate$A, mode="undirected")

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2))
plot(obj.igraph, vertex.color=karate$label, vertex.label=NA, 
     main="igraph visualization")
plot(obj.igraph, vertex.color=karate$label, vertex.label=NA,
     layout=lsm.embed$latent, main="Latent space embedding")
par(opar)
}

}
\concept{inference}
